#30 FIX : GameReady.tsx
